move: EpisodeStatus data 폴더로 이동
