+++ src/main/java/com/ham/netnovel/episode/service/EpisodeServiceImpl.java
@@ -7,7 +7,7 @@ import com.ham.netnovel.common.exception.ServiceMethodException;
 import com.ham.netnovel.common.message.RedisMessagePublisher;
 import com.ham.netnovel.episode.Episode;
 import com.ham.netnovel.episode.EpisodeRepository;
-import com.ham.netnovel.episode.EpisodeStatus;
+import com.ham.netnovel.episode.data.EpisodeStatus;
 import com.ham.netnovel.episode.dto.*;
 import com.ham.netnovel.novel.Novel;
 import com.ham.netnovel.novel.service.NovelService;
